Adds a cache for normalized rule sets
It has no impact on a long running application with one schema, but the
test suite takes 25% less time.
